+++ daemon/platform/linux/process_helper.rs
@@ -286,9 +286,12 @@ fn get_child_processes(pid: i32) -> Vec<Process> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use std::thread::sleep;
     use std::time::Duration;
+
+    use pretty_assertions::assert_eq;
+
+    use super::*;
 
     /// Assert that certain process id no longer exists
     fn process_is_gone(pid: i32) -> bool {
@@ -300,7 +303,7 @@ mod tests {
 
     #[test]
     fn test_spawn_command() {
-        let mut child = compile_shell_command("echo 'this is a test'")
+        let mut child = compile_shell_command("sleep 0.1")
             .spawn()
             .expect("Failed to spawn echo");